Make postconditions/throws sections consistent with existing condvarany wording.
Don't need to mention that postconditions are not honoured if an
exception is thrown as the remark already indicates that std::terminate()
is called if the postcondition fails. And exceptions thrown from pred()
cannot cause a postcondition failure.

Remove mention of std::bad_alloc in 'throws' sections as the stop_callback
constructor for registering a callback is now noexcept.